meta 不用protobuf

+++ 3rd/gbnet/buffer/compressed_def.h
@@ -1,6 +1,7 @@
 #pragma once
+#include "openssl\e_os2.h"
 
-enum CompressType : int
+enum CompressType : int8_t
 {
     CompressTypeNone = 0,
     CompressTypeGzip = 1,
